+++ Src/Public/Invoke-AsBuiltReport.Veeam.VBR.ps1
@@ -103,7 +103,28 @@ function Invoke-AsBuiltReport.Veeam.VBR {
         Get-AbrVbrServerConnection
         $VeeamBackupServer = ((Get-VBRServerSession).Server).ToString().ToUpper().Split(".")[0]
         $script:VbrLicenses = Get-VBRInstalledLicense
-        Section -Style Heading1 $($VeeamBackupServer) {
+
+        if ($Options.EnableDiagrams) {
+            Try {
+                Try {
+                    $Graph = Get-AbrVbrDiagrammer -DiagramType Backup-Infrastructure -DiagramOutput base64
+                } Catch {
+                    Write-PScriboMessage -IsWarning "Backup Infrastructure Diagram: $($_.Exception.Message)"
+                }
+                if ($Graph) {
+                    If ((Get-DiaImagePercent -GraphObj $Graph).Width -gt 1500) { $ImagePrty = 10 } else { $ImagePrty = 40 }
+                    Section -Style Heading1 "Backup Infrastructure Diagram." -Orientation Landscape {
+                        Image -Base64 $Graph -Text "Backup Infrastructure Diagram" -Percent $ImagePrty -Align Center
+                        Paragraph "Image preview: Opens the image in a new tab to view it at full resolution." -Tabs 2
+                    }
+                    BlankLine
+                }
+            } Catch {
+                Write-PScriboMessage -IsWarning "Backup Infrastructure Diagram Section: $($_.Exception.Message)"
+            }
+        }
+
+        Section -Style Heading1 $($VeeamBackupServer) -Orientation Portrait {
             Paragraph "The following section provides an overview of the implemented components of Veeam Backup & Replication."
             BlankLine
             #---------------------------------------------------------------------------------------------#